+++ ThesisApp/Utilities/Extensions/Double+ToString.swift
@@ -1,5 +1,5 @@
 //
-//  Float+ToString.swift
+//  Double+ToString.swift
 //  thesis-app
 //
 //  Created by Lisa Wittmann on 09.07.22.
@@ -7,12 +7,12 @@
 
 import Foundation
 
-extension Float {
+extension Double {
     
     func toString() -> String {
         let integer: Int = Int(self)
 
-        if (self - Float(integer)) > 0 {
+        if (self - Double(integer)) > 0 {
             return String(format: "%.2F", self)
         }
         return "\(integer)"
